# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/register/business": {
      "post": {
        "operationId": "AuthController_registerBusiness",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterBusinessDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/register/charity": {
      "post": {
        "operationId": "AuthController_registerCharity",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterCharityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/register/farmer-producer": {
      "post": {
        "operationId": "AuthController_registerFarmerProducer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterFarmerProducerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/register/farmer-consumer": {
      "post": {
        "operationId": "AuthController_registerFarmerConsumer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterFarmerConsumerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/register/platform-admin": {
      "post": {
        "operationId": "AuthController_registerPlatformAdmin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPlatformAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      },
      "patch": {
        "operationId": "AuthController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/resend-verification": {
      "post": {
        "operationId": "AuthController_sendVerificationOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerficationOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1/billing/trial": {
      "post": {
        "operationId": "BillingController_startTrial",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartTrialDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Starts the 30-day trial. Returns a Checkout URL — no card is required up\nfront. Collect payment details after the trial ends.",
        "tags": [
          "Billing"
        ]
      }
    },
    "/api/v1/billing/checkout": {
      "post": {
        "operationId": "BillingController_createCheckout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckoutSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Returns a Stripe-hosted Checkout URL for the client to redirect to.",
        "tags": [
          "Billing"
        ]
      }
    },
    "/api/v1/billing/change-plan": {
      "post": {
        "operationId": "BillingController_changePlan",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Switches an already-subscribed org between plans. Upgrades bill the\ndifference today; downgrades take effect at the period end.",
        "tags": [
          "Billing"
        ]
      }
    },
    "/api/v1/billing/change-plan/preview": {
      "post": {
        "operationId": "BillingController_previewChangePlan",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Dry run of `change-plan` — backs the confirmation dialog.",
        "tags": [
          "Billing"
        ]
      }
    },
    "/api/v1/billing/change-plan/pending": {
      "delete": {
        "operationId": "BillingController_cancelPendingChange",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Billing"
        ]
      }
    },
    "/api/v1/billing/portal": {
      "post": {
        "operationId": "BillingController_createPortal",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Stripe billing portal — card updates, invoices, self-serve cancellation.",
        "tags": [
          "Billing"
        ]
      }
    },
    "/api/v1/billing/cancel": {
      "post": {
        "operationId": "BillingController_cancel",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ends the plan at the close of the paid period — access continues until then.",
        "tags": [
          "Billing"
        ]
      }
    },
    "/api/v1/billing/resume": {
      "post": {
        "operationId": "BillingController_resume",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Undoes a scheduled cancellation before the period closes.",
        "tags": [
          "Billing"
        ]
      }
    },
    "/api/v1/billing/payments": {
      "get": {
        "operationId": "BillingController_listPayments",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Billing"
        ]
      }
    },
    "/api/v1/billing/enterprise-enquiry": {
      "post": {
        "operationId": "BillingController_enterpriseEnquiry",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnterpriseEnquiryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Billing"
        ]
      }
    },
...
  },
  "info": {
    "title": "Saveful for Business API",
    "description": "Food rescue platform API. All routes are prefixed `/api/v1`.\n\n**Authentication.** Every endpoint requires a bearer token unless marked otherwise. Obtain one from `POST /auth/login`, then send it as `Authorization: Bearer <token>`.\n\n**Enterprise endpoints** additionally require the organisation to be on the Enterprise plan, and are gated on the caller’s Enterprise role and scope. A 403 carrying `MISSING_PERMISSION` means the role is wrong; one carrying `OUTSIDE_SCOPE` means the role is right but the target sits outside what the caller can reach.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterBusinessDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "mobile": {
            "type": "string"
          },
          "businessName": {
            "type": "string"
          },
          "businessAddress": {
            "type": "string"
          },
          "registrationNumber": {
            "type": "string"
          },
          "brandName": {
            "type": "string"
          },
          "venueType": {
            "type": "string",
            "enum": [
              "CAFE_RESTAURANT",
              "BAKERY",
              "CATERER",
              "GROCERY_STORE",
              "FOOD_TRUCK",
              "CATERING_SERVICE",
              "HOTEL",
              "WEDDING_VENUE",
              "CLOUD_KITCHEN",
              "FARM",
              "PRODUCE_MARKET_GARDEN",
              "LIVESTOCK_FARM",
              "MIXED_FARM",
              "ORCHARD",
              "PROCESSING_FACILITY",
              "OTHER"
            ]
          },
          "orgType": {
            "type": "string",
            "enum": [
              "BUSINESS_SINGLE",
              "BUSINESS_MULTI",
              "CHARITY",
              "CHARITY_SINGLE",
              "CHARITY_MULTI",
              "FARMER_PRODUCER",
              "FARMER_CONSUMER"
            ]
          },
          "region": {
            "type": "string",
            "enum": [
              "IN",
              "US",
              "AU"
            ]
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password",
          "mobile",
          "businessName",
          "businessAddress",
          "orgType",
          "region",
          "latitude",
          "longitude"
        ]
      },
      "RegisterCharityDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "mobile": {
            "type": "string"
          },
          "charityName": {
            "type": "string"
          },
          "charityAddress": {
            "type": "string"
          },
          "registrationNumber": {
            "type": "string"
          },
          "brandName": {
            "type": "string"
          },
          "region": {
            "type": "string",
            "enum": [
              "IN",
              "US",
              "AU"
            ]
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "charityType": {
            "type": "string",
            "enum": [
              "BUSINESS_SINGLE",
              "BUSINESS_MULTI",
              "CHARITY",
              "CHARITY_SINGLE",
              "CHARITY_MULTI",
              "FARMER_PRODUCER",
              "FARMER_CONSUMER"
            ]
          },
          "pickupPostCode": {
            "type": "string"
          },
          "pickupRadiusKm": {
            "type": "number"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password",
          "charityName",
          "charityAddress",
          "region",
          "charityType"
        ]
      },
      "RegisterFarmerProducerDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "mobileNumber": {
            "type": "string"
          },
          "businessName": {
            "type": "string"
          },
          "businessAddress": {
            "type": "string"
          },
          "brandName": {
            "type": "string"
          },
          "venueType": {
            "type": "string",
            "enum": [
              "CAFE_RESTAURANT",
              "BAKERY",
              "CATERER",
              "GROCERY_STORE",
              "FOOD_TRUCK",
              "CATERING_SERVICE",
              "HOTEL",
              "WEDDING_VENUE",
              "CLOUD_KITCHEN",
              "FARM",
              "PRODUCE_MARKET_GARDEN",
              "LIVESTOCK_FARM",
              "MIXED_FARM",
              "ORCHARD",
              "PROCESSING_FACILITY",
              "OTHER"
            ]
          },
          "orgType": {
            "type": "string",
            "enum": [
              "BUSINESS_SINGLE",
              "BUSINESS_MULTI",
              "CHARITY",
              "CHARITY_SINGLE",
              "CHARITY_MULTI",
              "FARMER_PRODUCER",
              "FARMER_CONSUMER"
            ]
          },
          "region": {
            "type": "string",
            "enum": [
              "IN",
              "US",
              "AU"
            ]
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password",
          "businessName",
          "businessAddress",
          "orgType",
          "region",
          "latitude",
          "longitude"
        ]
      },
      "RegisterFarmerConsumerDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "mobile": {
            "type": "string"
          },
          "farmName": {
            "type": "string"
          },
          "businessName": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "brandName": {
            "type": "string"
          },
          "venueType": {
            "type": "string",
            "enum": [
              "CAFE_RESTAURANT",
              "BAKERY",
              "CATERER",
              "GROCERY_STORE",
              "FOOD_TRUCK",
              "CATERING_SERVICE",
              "HOTEL",
              "WEDDING_VENUE",
              "CLOUD_KITCHEN",
              "FARM",
              "PRODUCE_MARKET_GARDEN",
              "LIVESTOCK_FARM",
              "MIXED_FARM",
              "ORCHARD",
              "PROCESSING_FACILITY",
              "OTHER"
            ]
          },
          "region": {
            "type": "string",
            "enum": [
              "IN",
              "US",
              "AU"
            ]
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password",
          "farmName",
          "businessName",
          "address",
          "region",
          "latitude",
          "longitude"
        ]
      },
      "RegisterPlatformAdminDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "VerifyEmailOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "otp": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "otp"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string"
          }
        },
        "required": [
          "phoneNumber"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "otp": {
            "type": "string"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "otp",
          "newPassword"
        ]
      },
      "ResendVerficationOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
...
      "StartTrialDto": {
        "type": "object",
        "properties": {
          "planId": {
            "type": "number"
          },
          "billingCycle": {
            "type": "string",
            "description": "Which cycle to charge once the trial converts. Defaults to monthly.",
            "enum": [
              "MONTHLY",
              "ANNUAL"
            ]
          }
        },
        "required": [
          "planId"
        ]
      },
      "CreateCheckoutSessionDto": {
        "type": "object",
        "properties": {
          "planId": {
            "type": "number"
          },
          "billingCycle": {
            "type": "string",
            "enum": [
              "MONTHLY",
              "ANNUAL"
            ]
          }
        },
        "required": [
          "planId",
          "billingCycle"
        ]
      },
      "ChangePlanDto": {
        "type": "object",
        "properties": {
          "planId": {
            "type": "number"
          },
          "billingCycle": {
            "description": "Omit to keep the current cycle.",
            "enum": [
              "MONTHLY",
              "ANNUAL"
            ],
            "type": "string"
          }
        },
        "required": [
          "planId"
        ]
      },
      "EnterpriseEnquiryDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "businessName": {
            "type": "string"
          },
          "businessType": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "locationBand": {
            "type": "string",
            "enum": [
              "BAND_10_25",
              "BAND_26_50",
              "BAND_51_100",
              "BAND_100_PLUS"
            ]
          },
          "contactWindow": {
            "type": "string",
            "enum": [
              "ASAP",
              "MORNING",
              "AFTERNOON",
              "ANY_TIME"
            ]
          },
          "message": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": [
          "firstName",
          "lastName",
          "businessName",
          "businessType",
          "mobile",
          "locationBand",
          "contactWindow"
        ]
      },
...
    }
  }
}